feat(go): register user

+++ go/internal/handler/interface.go
@@ -6,6 +6,6 @@ import (
 )
 
 type UserService interface {
-	Register(ctx context.Context, data service.UserServiceRegisterInput) (int, error)
-	CreateJwtToken(ctx context.Context, userId int, email string) (string, error)
+	Register(ctx context.Context, data service.UserServiceRegisterInput) (int64, error)
+	CreateJwtToken(ctx context.Context, userId int64, email string) (string, error)
 }
